+++ helpers/test.js
@@ -175,6 +175,13 @@ test('formatHttpRequest and formatHttpResponse should return a valid ecs object'
         body: { bytes: contentLen }
       }
     })
+    // https://www.elastic.co/guide/en/ecs/current/ecs-client.html fields
+    t.ok(line.client, 'client fields are set')
+    t.ok(line.client.address === '127.0.0.1' || line.client.address === '::ffff:127.0.0.1',
+      'client.address is set')
+    t.ok(line.client.ip === line.client.address,
+      'client.address duplicated to client.ip')
+    t.equal(typeof (line.client.port), 'number')
 
     res.end(resBody)
   }